Fix test (false positive)

+++ dumusicfile/dutrack.h
@@ -23,7 +23,7 @@ public:
 
     static DuTrackPtr fromDuMusicBinary(const music_track &du_track,
                                         const music_sample *du_sample_start,
-                                        int totalSample);
+                                        uint16_t totalSample);
     static DuTrackPtr fromDuMusicBinary(const music_track &du_track);
     static DuTrackPtr fromJson(const QJsonObject &jsonTrack);
